requirements.txt
alembic==1.17.0
aiokafka==0.12.0
asyncpg==0.30.0
confluent-kafka==2.11.1
faststream==0.6.1
psycopg2-binary==2.9.11
pydantic==2.11.5
pydantic-settings==2.9.1
SQLAlchemy==2.0.44
faust-streaming==0.11.3
fastapi==0.119.0
hdfs==2.7.3